social-flow-backend/testing/e2e/selenium/requirements.txt
selenium==4.19.0
pytest==7.4.0
pytest-xdist==3.2.0
webdriver-manager==4.11.0
python-dotenv==1.0.0
pytest-html==3.2.0
allure-pytest==2.11.1
requests==2.31.0
